Use a non-quadratic collection frequency.
This is still very naive, butavoids quadratic behaviour, by collecting
exponentially rarely as the heap grows. This cuts the run time of the
included `keep_10000` benchmark from an (estimated) 300 ms to ~1ms.

The failure case of the naive strategy is how often it collects when
growing to some heap size N (e.g. a graph structure with N nodes): if
the N is larger than the threshold, then at some point the live heap
size will have passed the threshold and every allocation will be trying
to collect. Each collection does O(size of heap) work as it scans the
heap, and each allocation is linearly increasing the size of heap, so to
get to a heap size of N, we do O(N) collections that involve O(N) work:
O(N^2).

By collecting exponentially rarely, getting to a heap size of N takes
one collection of size O(N), one of size O(N r), one of size O(N
r^2) (for some r < 1), which is O(N) in total. (I chose r somewhat
randomly to be 0.7 in this patch.)