Update README.md
I've changed two instances of struct fields whose type is a garbage-collected handle to the struct type itself. It seemed unlikely that there was a safe way to create instances of the type, a point which caused some confusion among readers. The type was changed to `Option<Gc<Foo>>`, which is I think is the right call, though it could also have been `Gc<Option<Foo>>`; I don't have much experience with the library, and if you think the second type is more idiomatic let me know.