Gc::from_raw: Rely only on documented guarantees to compute layout
The claim “Because it is ?Sized, it will always be the last field in
memory” was not reliable: the ?Sized type may in fact be Sized (which
the compiler will know after monomorphization), and in any case, the
language provides no guarantees about the layout of repr(Rust)
structs.

Switch GcBox to repr(C), and rewrite the layout computation based on
Layout::extend, which has a documented guarantee to match the layout
of repr(C) structs.

(RcBox in the standard library is repr(C) for the same reason.)

Signed-off-by: Anders Kaseorg <andersk@mit.edu>